修正bug：删除商品分类的时候没有判断子分类是否存在

+++ web/Application/Home/Service/GoodsService.class.php
@@ -558,6 +558,16 @@ class GoodsService extends PSIBaseService {
 			return $this->bad("还有属于商品分类 [{$name}] 的商品，不能删除该分类");
 		}
 		
+		// 判断是否还有子分类
+		$sql = "select count(*) as cnt from t_goods_category 
+				where parent_id = '%s' ";
+		$data = $db->query($sql, $id);
+		$cnt = $data[0]["cnt"];
+		if ($cnt > 0) {
+			$db->rollback();
+			return $this->bad("分类[{$name}]还有子分类，不能删除");
+		}
+		
 		$sql = "delete from t_goods_category where id = '%s' ";
 		$rc = $db->execute($sql, $id);
 		if ($rc === false) {
